Split up data-explorer filters
Into separate files
