+++ app/data-explorer/page.tsx
@@ -28,13 +28,6 @@ import {
   ColumnsEffects,
 } from "@/components/data-explorer/table-columns"
 import { DataTable } from "@/components/data-explorer/table"
-import {
-  FilterOutcomes,
-  FilterInterventions,
-  FilterPapers,
-  FilterStudies,
-  FilterEffects,
-} from "@/components/data-explorer/filters"
 
 import papers from "@/assets/data/papers.json"
 import studies from "@/assets/data/studies.json"
@@ -44,6 +37,11 @@ import effects from "@/assets/data/effects.json"
 import all from "@/assets/data/data.json"
 
 import { semiJoin } from "@/lib/json-functions"
+import { FilterPapers } from "@/components/data-explorer/filters/filter-papers"
+import { FilterStudies } from "@/components/data-explorer/filters/filter-studies"
+import { FilterInterventions } from "@/components/data-explorer/filters/filter-interventions"
+import { FilterOutcomes } from "@/components/data-explorer/filters/filter-outcomes"
+import { FilterEffects } from "@/components/data-explorer/filters/filter-effects"
 
 export default function DataExplorer() {
   const [level, setLevel] = useState("paper")
